# contracts/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evidrilo API",
    "version": "1.0.0",
    "description": "Public contract for the versioned Evidrilo API routes."
  },
  "paths": {
...
    "/v1/project-ai/activity": {
      "get": {
        "operationId": "get_v1_project_ai_activity",
        "tags": [
          "project-ai"
        ],
        "security": [
          {
            "SupabaseBearer": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/project-ai-activity-history.v1"
                }
              }
            }
          },
          "default": {
            "description": "API error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/http-errors.v1"
                }
              }
            }
          }
        },
        "x-evidrilo-test": {
          "file": "platform/api.Tests/ProjectAiScaffoldFlowTests.cs",
          "anchor": "/v1/project-ai/activity"
        },
        "parameters": [
          {
            "name": "installationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "projectId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of metadata records returned; defaults to 50.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque cursor returned by the previous page.",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "parameters": []
    },
    "/v1/project-ai/activity/general": {
      "delete": {
        "operationId": "delete_v1_project_ai_activity_general",
        "tags": [
          "project-ai"
        ],
        "security": [
          {
            "SupabaseBearer": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/project-ai-activity-clear.v1"
                }
              }
            }
          },
          "default": {
            "description": "API error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/http-errors.v1"
                }
              }
            }
          }
        },
        "x-evidrilo-test": {
          "file": "platform/api.Tests/ProjectAiScaffoldFlowTests.cs",
          "anchor": "/v1/project-ai/activity/general"
        },
        "parameters": [
          {
            "name": "installationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "parameters": []
    },
...
  },
  "components": {
    "securitySchemes": {
      "SupabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "RevenueCatWebhookAuthorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      },
      "RevenueCatWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RevenueCat-Webhook-Signature"
      }
    },
    "schemas": {
...
      "http-errors.v1": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/http-errors.v1.json",
        "title": "Evidrilo HTTP error response v1",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.http-error"
          },
          "version": {
            "const": "1"
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{2,63}$"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 240
          },
          "requestId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9_-]+$"
          },
          "creditCost": {
            "description": "Present only when verified provider usage was charged even though the requested response was withheld.",
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          }
        }
      },
...
      "project-ai-activity-clear.v1": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/project-ai-activity-clear.v1.json",
        "title": "Evidrilo General chat activity clear response v1",
        "description": "Clears unlinked metadata for one authenticated account and installation; it never mutates projects.",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "clearedCount"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.project-ai-activity"
          },
          "version": {
            "const": "1"
          },
          "clearedCount": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "project-ai-activity-history.v1": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/project-ai-activity-history.v1.json",
        "title": "Evidrilo per-installation Project AI activity history v1",
        "description": "Metadata only. General entries are unlinked. Prompt and response transcripts are not included or retained by this contract.",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "activities",
          "nextCursor"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.project-ai-activity"
          },
          "version": {
            "const": "1"
          },
          "activities": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "activityId",
                "mode",
                "projectId",
                "stageId",
                "operationId",
                "baseProjectRevision",
                "resultProjectRevision",
                "outcome",
                "createdAt",
                "updatedAt"
              ],
              "properties": {
                "activityId": {
                  "type": "string",
                  "format": "uuid"
                },
                "mode": {
                  "enum": [
                    "PROJECT",
                    "GENERAL"
                  ]
                },
                "projectId": {
                  "oneOf": [
                    {
                      "type": "string",
                      "format": "uuid"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "stageId": {
                  "oneOf": [
                    {
                      "type": "string",
                      "maxLength": 80
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "operationId": {
                  "oneOf": [
                    {
                      "type": "string",
                      "maxLength": 80
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "baseProjectRevision": {
                  "oneOf": [
                    {
                      "type": "integer",
                      "minimum": 1
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "resultProjectRevision": {
                  "oneOf": [
                    {
                      "type": "integer",
                      "minimum": 1
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "outcome": {
                  "enum": [
                    "PENDING",
                    "APPLIED",
                    "EDITED",
                    "DISMISSED",
                    "STALE",
                    "FAILED",
                    "COMPLETED"
                  ]
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "nextCursor": {
            "oneOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
    }
  }
}